moved IpEndpoint to main library
