+++ include/jmg/ip_endpoint.h
